requirements/test.txt
# Dependencies related to testing stagegage_api
-r base.txt

# Mocking
mock==1.0.1

# Fixtures
factory-boy==2.5.2
fake-factory==0.5.2

# Test Runner
django-nose==1.4.1
nose-progressive==1.5.1

